[feat/#8] 패키지 분리
