+++ app/src/main/java/com/sopt/dive/data/remote/dto/request/SignUpRequestDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.dive.data.dto.request
+package com.sopt.dive.data.remote.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
