+++ app/src/main/java/com/sopt/dive/data/remote/mapper/LoginResponseMapper.kt
@@ -1,6 +1,6 @@
-package com.sopt.dive.data.mapper
+package com.sopt.dive.data.remote.mapper
 
-import com.sopt.dive.data.dto.response.LoginResponseDto
+import com.sopt.dive.data.remote.dto.response.LoginResponseDto
 import com.sopt.dive.domain.model.LoginModel
 
 fun LoginResponseDto.toDomain(): LoginModel {
