+++ app/src/main/java/com/sopt/dive/data/remote/repositoryimpl/AuthRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.sopt.dive.data.repositoryimpl
+package com.sopt.dive.data.remote.repositoryimpl
 
-import com.sopt.dive.data.datasource.AuthDataSource
-import com.sopt.dive.data.dto.request.LoginRequestDto
-import com.sopt.dive.data.mapper.toDomain
+import com.sopt.dive.data.remote.datasource.AuthDataSource
+import com.sopt.dive.data.remote.dto.request.LoginRequestDto
+import com.sopt.dive.data.remote.mapper.toDomain
 import com.sopt.dive.data.util.handleApiResponse
 import com.sopt.dive.data.util.safeApiCall
 import com.sopt.dive.domain.model.LoginModel
