+++ app/src/main/java/com/sopt/dive/domain/repository/AuthRepository.kt
@@ -1,6 +1,6 @@
 package com.sopt.dive.domain.repository
 
-import com.sopt.dive.data.dto.request.LoginRequestDto
+import com.sopt.dive.data.remote.dto.request.LoginRequestDto
 import com.sopt.dive.domain.model.LoginModel
 
 interface AuthRepository {
